# Conflicts:
#	jayud-ocean-ship/jayud-ocean-ship-web/src/main/java/com/jayud/oceanship/controller/SeaOrderController.java

+++ jayud-ocean-ship/jayud-ocean-ship-service/src/main/java/com/jayud/oceanship/service/impl/SeaOrderServiceImpl.java
@@ -75,6 +75,12 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
     @Autowired
     private ISeaContainerInformationService seaContainerInformationService;
 
+    @Autowired
+    private ISeaPortService seaPortService;
+
+    @Autowired
+    private ICabinetTypeService cabinetTypeService;
+
     @Override
     @Transactional
     public String createOrder(AddSeaOrderForm addSeaOrderForm) {
@@ -99,13 +105,10 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
             this.updateById(seaOrder);
         }
         //获取柜型数量
+        //先删除原来的柜型
+        this.cabinetSizeNumberService.deleteCabinet(addSeaOrderForm.getOrderId());
+
         if(addSeaOrderForm.getCabinetType()!=null){
-            if (addSeaOrderForm.getCabinetType().equals(2)) {
-                //先删除原来的柜型
-
-                this.cabinetSizeNumberService.deleteCabinet(addSeaOrderForm.getOrderId());
-
-            }
             if (addSeaOrderForm.getCabinetType().equals(1)) {
                 List<CabinetSizeNumber> cabinetSizeNumbers = addSeaOrderForm.getCabinetSizeNumbers();
 
@@ -213,13 +216,13 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
         //查询商品信息
         ApiResult<List<GoodsVO>> result = this.omsClient.getGoodsByBusOrders(Collections.singletonList(seaOrderVO.getOrderNo()), businessType);
         if (result.getCode() != HttpStatus.SC_OK) {
-            log.warn("查询商品信息失败 airOrderId={}");
+            log.warn("查询商品信息失败 seaOrderId={}");
         }
         seaOrderVO.setGoodsForms(result.getData());
         //查询地址信息
         ApiResult<List<OrderAddressVO>> resultOne = this.omsClient.getOrderAddressByBusOrders(Collections.singletonList(seaOrderVO.getOrderNo()), businessType);
         if (resultOne.getCode() != HttpStatus.SC_OK) {
-            log.warn("查询订单地址信息失败 airOrderId={}");
+            log.warn("查询订单地址信息失败 seaOrderId={}");
         }
         //处理地址信息
         for (OrderAddressVO address : resultOne.getData()) {
@@ -391,14 +394,22 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
     public void updateOrSaveProcessStatus(SeaProcessOptForm form) {
 
         //删除补料信息
-        seaReplenishmentService.deleteSeaReplenishment(form.getOrderId(),form.getOrderNo());
         List<SeaReplenishment> list = seaReplenishmentService.getList(form.getOrderId(),form.getOrderNo());
         List<String> orderNo = new ArrayList<>();
         for (SeaReplenishment replenishment : list) {
             orderNo.add(replenishment.getOrderNo());
         }
-        omsClient.deleteOrderAddressByBusOrders(orderNo,BusinessTypeEnum.HY.getCode());
-        omsClient.deleteGoodsByBusOrders(orderNo,BusinessTypeEnum.HY.getCode());
+        seaReplenishmentService.deleteSeaReplenishment(form.getOrderId(),form.getOrderNo());
+
+        if(orderNo.size()>0){
+            omsClient.deleteOrderAddressByBusOrders(orderNo,BusinessTypeEnum.HY.getCode());
+            omsClient.deleteGoodsByBusOrders(orderNo,BusinessTypeEnum.HY.getCode());
+            int i1 = seaContainerInformationService.deleteSeaContainerInfo(orderNo);
+            if(i1<=0){
+                log.warn("货柜信息添加失败");
+            }
+        }
+
 
         if (form.getType().equals(1)) {//合并，多个订单合并成一个补料
             List<AddSeaOrderForm> seaOrderForms = form.getSeaOrderForms();
@@ -429,6 +440,7 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
             replenishment.setOrderNo(blOrderNo);
             replenishment.setIsBillOfLading(0);
             replenishment.setIsReleaseOrder(0);
+            replenishment.setCabinetTypeName(cabinetTypeService.getCabinetTypeName(replenishment.getCabinetType()));
             boolean save = seaReplenishmentService.save(replenishment);
             if (!save) {
                 log.warn("合并补料信息添加失败");
@@ -499,6 +511,7 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
                 replenishment.setOrderNo(getBLOrderNo(seaOrderForms.get(0).getOrderNo(), form.getType(), seaReplenishments.size(), i + 1));
                 replenishment.setIsBillOfLading(0);
                 replenishment.setIsReleaseOrder(0);
+                replenishment.setCabinetTypeName(cabinetTypeService.getCabinetTypeName(replenishment.getCabinetType()));
                 boolean save = seaReplenishmentService.save(replenishment);
                 if (!save) {
                     log.warn("分单补料信息添加失败");
@@ -667,8 +680,6 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
         List<SeaReplenishmentVO> seaReplenishmentVOS = ConvertUtil.convertList(list, SeaReplenishmentVO.class);
         for (SeaReplenishmentVO seaReplenishmentVO : seaReplenishmentVOS) {
             //获取截补料中的柜型数量以及货柜信息
-            List<CabinetSizeNumberVO> list1 = cabinetSizeNumberService.getList(seaReplenishmentVO.getId());
-            seaReplenishmentVO.setCabinetSizeNumbers(list1);
             List<SeaContainerInformationVO> seaContainerInformations = seaContainerInformationService.getList(seaReplenishmentVO.getId());
             seaReplenishmentVO.setSeaContainerInformations(seaContainerInformations);
 
@@ -688,6 +699,9 @@ public class SeaOrderServiceImpl extends ServiceImpl<SeaOrderMapper, SeaOrder> i
                 address.getFile(prePath);
                 seaReplenishmentVO.processingAddress(address);
             }
+            //匹配港口信息
+            seaReplenishmentVO.setPortDepartureName(seaPortService.getPortName(seaReplenishmentVO.getPortDepartureCode()));
+            seaReplenishmentVO.setPortDestinationName(seaPortService.getPortName(seaReplenishmentVO.getPortDestinationCode()));
         }
         seaOrder.setSeaReplenishments(seaReplenishmentVOS);
         return seaOrder;
